+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Added a check for `requires_sudo` to `run_all_tasks()` and `main()`.
+  - If `requires_sudo` is `true` in `run_all_tasks()`, the task is skipped. In `main()`, the user is prompted to rerun command with sudo.
 
 ## [0.6.1] - 2025-10-22